Inheritance-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

web3==6.15.1
python-dotenv==1.0.1

requests==2.31.0
cryptography==42.0.5

pydantic==2.6.1

python-multipart==0.0.9
